+++ actix-server/src/lib.rs
@@ -11,6 +11,7 @@ mod server;
 mod service;
 mod signals;
 mod socket;
+mod test_server;
 mod waker_queue;
 mod worker;
 
@@ -18,6 +19,7 @@ pub use self::builder::ServerBuilder;
 pub use self::config::{ServiceConfig, ServiceRuntime};
 pub use self::server::Server;
 pub use self::service::ServiceFactory;
+pub use self::test_server::TestServer;
 
 #[doc(hidden)]
 pub use self::socket::FromStream;
@@ -123,5 +125,19 @@ impl<T> Future for JoinAll<T> {
         } else {
             Poll::Pending
         }
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+
+    #[actix_rt::test]
+    async fn test_join_all() {
+        let futs = vec![ready(Ok(1)), ready(Err(3)), ready(Ok(9))];
+        let mut res = join_all(futs).await.into_iter();
+        assert_eq!(Ok(1), res.next().unwrap());
+        assert_eq!(Err(3), res.next().unwrap());
+        assert_eq!(Ok(9), res.next().unwrap());
     }
 }